feat: adjust dependenciesplugin to buildSrc
