+++ android/buildSrc/src/main/java/br/com/zup/beagle/Dependencies.kt
