+++ backend/backend-widgets/build.gradle
@@ -27,4 +27,4 @@ dependencies {
 
 }
 
-//apply from: rootProject.file('maven-publish.gradle')
+apply from: rootProject.file('../maven-publish.gradle')
