+++ backend/build.gradle
@@ -24,6 +24,8 @@ buildscript {
         classpath "com.android.tools.build:gradle:4.0.1"
         classpath "org.jetbrains.kotlin:kotlin-gradle-plugin:1.3.72"
         classpath "org.jetbrains.dokka:dokka-gradle-plugin:0.10.1"
+        //TODO remove this line after migrate to mavenCentral
+        classpath "com.github.dcendents:android-maven-gradle-plugin:2.1"
     }
 }
