+++ backend/buildSrc/src/main/java/br/com/zup/beagle/Dependencies.kt
@@ -1,0 +1,121 @@
+/*
+ * Copyright 2020 ZUP IT SERVICOS EM TECNOLOGIA E INOVACAO SA
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+
+package br.com.zup.beagle
+
+import org.gradle.api.Plugin
+import org.gradle.api.Project
+
+class Dependencies : Plugin<Project> {
+    override fun apply(project: Project) {}
+
+    object ApplicationId {
+        const val id = "br.com.zup.beagle.sample"
+        const val micronautSample = "br.com.zup.beagle.sample.micronaut.BeagleUiSampleApplication"
+    }
+
+    object GroupId {
+        const val backendSample = "br.com.zup.beagle.sample"
+        const val backendStarters = "br.com.zup.beagle.spring-starter"
+    }
+
+    object Modules {
+        const val beagleBackendWidgets = ":backend-widgets"
+        const val beagleBackendProcessor = ":processor"
+        const val beagleBackendFramework = ":framework"
+
+        const val backendSampleCore = ":sample:core"
+        const val backendSampleSpring = ":sample:spring"
+        const val backendSampleMicronaut = ":sample:micronaut"
+
+        const val beagleBackendSpringStarter = ":starters:beagle-spring-starter"
+        const val beagleBackendMicronautStarter = ":starters:beagle-micronaut-starter"
+
+        const val beagleSchemaKotlinCore = ":schema:kotlin-core"
+
+        const val beagleCommonAnnotation = ":common:annotation"
+        const val beagleCommonProcessorUtils = ":common:processor-utils"
+        const val beagleExtendedSchema = ":common:extended-schema"
+    }
+
+    object Releases {
+        const val beagleVersionName = "0.0.1-SNAPSHOT"
+    }
+
+    object Versions {
+        const val java = "1.8"
+        const val kotlin = "1.3.72"
+
+        const val kotlinPoet = "1.5.0"
+
+        const val junit5 = "5.6.2"
+
+        const val kotlinTest = "1.3.72"
+        const val googleAutoService = "1.0-rc6"
+
+        const val mockk = "1.10.0"
+
+        const val jacksonKotlin = "2.11.1"
+
+        const val guava = "29.0-jre"
+        const val slf4j = "1.7.30"
+
+        const val springBoot = "2.3.1.RELEASE"
+        const val micronaut = "2.0.0"
+
+        const val incap = "0.2"
+    }
+
+    object BackendLibraries {
+        const val guava = "com.google.guava:guava:${Versions.guava}"
+        const val slf4jJUL = "org.slf4j:slf4j-jdk14:${Versions.slf4j}"
+    }
+
+    object ProcessorLibraries {
+        const val autoService = "com.google.auto.service:auto-service:${Versions.googleAutoService}"
+        const val incap = "net.ltgt.gradle.incap:incap:${Versions.incap}"
+        const val incapPrcessor = "net.ltgt.gradle.incap:incap-processor:${Versions.incap}"
+    }
+
+    object GeneralLibraries {
+        const val kotlin = "org.jetbrains.kotlin:kotlin-stdlib-jdk8:${Versions.kotlin}"
+        const val kotlinReflect = "org.jetbrains.kotlin:kotlin-reflect:${Versions.kotlin}"
+        const val kotlinPoet = "com.squareup:kotlinpoet:${Versions.kotlinPoet}"
+        const val jacksonKotlin = "com.fasterxml.jackson.module:jackson-module-kotlin:${Versions.jacksonKotlin}"
+    }
+
+    object TestLibraries {
+        const val junit5Api = "org.junit.jupiter:junit-jupiter-api:${Versions.junit5}"
+        const val junit5Engine = "org.junit.jupiter:junit-jupiter-engine:${Versions.junit5}"
+        const val mockk = "io.mockk:mockk:${Versions.mockk}"
+        const val kotlinTest = "org.jetbrains.kotlin:kotlin-test:${Versions.kotlinTest}"
+    }
+
+    object SpringLibraries {
+        const val autoconfigure = "org.springframework.boot:spring-boot-autoconfigure:${Versions.springBoot}"
+        const val autoconfigureProcessor = "org.springframework.boot:spring-boot-autoconfigure-processor:${Versions.springBoot}"
+        const val webStarter = "org.springframework.boot:spring-boot-starter-web:${Versions.springBoot}"
+        const val testStarter = "org.springframework.boot:spring-boot-starter-test:${Versions.springBoot}"
+    }
+
+    object MicronautLibraries {
+        const val runtime = "io.micronaut:micronaut-runtime:${Versions.micronaut}"
+        const val netty = "io.micronaut:micronaut-http-server-netty:${Versions.micronaut}"
+        const val injectJava = "io.micronaut:micronaut-inject-java:${Versions.micronaut}"
+    }
+
+}
