+++ backend/buildSrc/src/main/resources/META-INF/gradle-plugin/br.com.zup.beagle.dependencies.properties
@@ -1,0 +1,1 @@
+implementation-class=br.com.zup.beagle.Dependencies
