--- backend/dependenciesPlugin/build.gradle
@@ -1,29 +1,0 @@
-buildscript {
-    repositories {
-        jcenter()
-    }
-    dependencies {
-        classpath "org.jetbrains.kotlin:kotlin-gradle-plugin:1.3.72"
-    }
-}
-
-apply plugin: 'kotlin'
-apply plugin: 'java-gradle-plugin'
-
-repositories {
-    jcenter()
-}
-
-dependencies {
-    implementation gradleApi()
-    implementation "org.jetbrains.kotlin:kotlin-stdlib-jdk8:1.3.72"
-}
-
-gradlePlugin {
-    plugins {
-        dependencies {
-            id = 'br.com.zup.beagle.dependencies'
-            implementationClass = 'br.com.zup.beagle.Dependencies'
-        }
-    }
-}
