+++ backend/framework/build.gradle
@@ -23,7 +23,7 @@ dependencies {
     implementation Dependencies.GeneralLibraries.kotlin
     implementation Dependencies.GeneralLibraries.jacksonKotlin
     implementation Dependencies.BackendLibraries.guava
-    api project(Dependencies.Modules.beagleSchemaBackendWidgets)
+    api project(Dependencies.Modules.beagleBackendWidgets)
     api project(Dependencies.Modules.beagleCommonAnnotation)
     api project(Dependencies.Modules.beagleSchemaKotlinCore)
     api project(Dependencies.Modules.beagleExtendedSchema)
@@ -37,4 +37,4 @@ test {
     useJUnitPlatform()
 }
 
-//apply from: rootProject.file('maven-publish.gradle')
+apply from: rootProject.file('../maven-publish.gradle')
