+++ backend/starters/beagle-micronaut-starter/build.gradle
@@ -47,4 +47,4 @@ compileTestKotlin {
     kotlinOptions.jvmTarget = Dependencies.Versions.java
 }
 
-//apply from: rootProject.file('maven-publish.gradle')
+apply from: rootProject.file('../maven-publish.gradle')
