+++ common/extended-schema/build.gradle
@@ -28,4 +28,4 @@ dependencies {
     testRuntime Dependencies.TestLibraries.junit5Engine
 }
 
-// apply from: rootProject.file('../maven-publish.gradle')
+ apply from: rootProject.file('../maven-publish.gradle')
